mm-video-v4l2: vidc: venc: Optimize color align function
For non-mod 16 resolutions memmove is taking long time. This
may be due to the implementation of memmove. To enhance the
performance, always memmove the pixels in 16 aligned blocks.
FW ignores the padding in YUV anyways.

Also allocate cache buffers so that color align is faster.

These improve the color conversion performance significantly.

Change-Id: I6569011aa35a4d3f621628e9009e21479c30f214
CRs-Fixed: 927541