mm-video: refine condition for free buffer call
When there is a failure during transition from state loaded
to state idle, client requests to free the allocated buffers.
As per present conditional checks it is not allowed, to free
resources in loaded state which leads to memory leak for
previously allocated resources. Refining conditions to allow
free buffer call to address this.

Change-Id: I0ec58de48adcbf75179ac79433c6839309413243
Signed-off-by: Aniket Masule <amasule@codeaurora.org>